move getAdmin() to ContainerCluster
